+++ src/tasks/mocha/__tests__/cli-options/gulpfile.js
@@ -4,4 +4,6 @@ const gulp = require('gulp');
 const task = require('../../index');
 
 task.dependencies = {};
-task.register(gulp, 'mocha', true);
+task.register(gulp, 'mocha', {
+  files: 'test/index-test.js'
+});