+++ src/tasks/mocha/defaults.js
@@ -4,6 +4,17 @@ const path = require('path');
 const processCwd = process.cwd();
 
 module.exports = {
-  files: path.join(processCwd, 'src/test/**/*.+(js|ts|tsx)'),
-  require: `${path.join(__dirname, 'setup-file.js')}`
+  files: path.join(processCwd, 'src/+(test|tests)/**/*.+(js|ts|tsx)'),
+  require: `${path.join(__dirname, 'setup-file.js')}`,
+  maxConcurrency: 0,
+  collectCoverage: true,
+  coverageReporters: ['text-summary', 'html'],
+  coverageDirectory: 'coverage',
+  coverageFrom: [
+    path.join(processCwd, 'src/**/*.+(js|ts|tsx)'),
+    `!${path.join(processCwd, 'src/**/*-+(spec|test).+(js|ts|tsx)')}`,
+    `!${path.join(processCwd, 'src/**/__tests__/**/*')}`,
+    `!${path.join(processCwd, 'src/**/+(test|tests)/**/*')}`
+  ],
+  coverageThresholds: null
 };